#3606 [iOS] add tap anywhere / scroll / swipe right to cancel delete button

+++ iOS/SmartDeviceApp/Classes/Views/CollectionViewCells/PrintJobHistoryGroupCell.m
@@ -29,7 +29,10 @@
 @property (weak, nonatomic) IBOutlet UIButton* groupName;
 @property (weak, nonatomic) IBOutlet UIButton* groupIndicator;
 @property (weak, nonatomic) IBOutlet UIButton* deleteAllButton;
-@property (weak, nonatomic) IBOutlet UITableView* tableView;
+@property (weak, nonatomic) IBOutlet UITableView* itemsView;
+
+/** Keeps trach of the index of an item that has the delete button. */
+@property (strong, nonatomic) NSIndexPath* deleteItem;
 
 #pragma mark - Data Properties
 
@@ -75,34 +78,37 @@
 
 - (UITableViewCell*)tableView:(UITableView*)tableView cellForRowAtIndexPath:(NSIndexPath*)indexPath
 {
-    UITableViewCell* cell = [tableView dequeueReusableCellWithIdentifier:ITEMCELL
-                                                            forIndexPath:indexPath];
+    UITableViewCell* itemCell = [tableView dequeueReusableCellWithIdentifier:ITEMCELL
+                                                                forIndexPath:indexPath];
     
     // get print job details (name, result, timestamp)
     NSArray* printJob = [self.listPrintJobs objectAtIndex:indexPath.row];
     
     // print job name
-    cell.textLabel.text = [NSString stringWithFormat:@"%@", printJob[IDX_NAME]];
+    itemCell.textLabel.text = [NSString stringWithFormat:@"%@", printJob[IDX_NAME]];
     
     // print job result
     BOOL result = [printJob[IDX_RESULT] boolValue];
     if (result)
-        cell.imageView.image = [UIImage imageNamed:IMAGE_JOB_STATUS_OK];
+        itemCell.imageView.image = [UIImage imageNamed:IMAGE_JOB_STATUS_OK];
     else
-        cell.imageView.image = [UIImage imageNamed:IMAGE_JOB_STATUS_NG];
+        itemCell.imageView.image = [UIImage imageNamed:IMAGE_JOB_STATUS_NG];
     
     // print job timestamp
     NSDateFormatter* timestampFormat = [[NSDateFormatter alloc] init];
     [timestampFormat setDateFormat:FORMAT_DATE_TIME];
     [timestampFormat setTimeZone:[NSTimeZone timeZoneWithName:FORMAT_ZONE]]; //TODO: should depend on localization?
-    cell.detailTextLabel.text = [timestampFormat stringFromDate:printJob[IDX_TIMESTAMP]];
-    cell.detailTextLabel.hidden = NO;
-    cell.accessoryView = nil;
+    itemCell.detailTextLabel.text = [timestampFormat stringFromDate:printJob[IDX_TIMESTAMP]];
+    itemCell.detailTextLabel.hidden = NO;
+    itemCell.accessoryView = nil;
     
     // fix for the bugged always-white cell in iPad iOS7
-    cell.backgroundColor = [UIColor clearColor];
+    itemCell.backgroundColor = [UIColor clearColor];
     
-    return cell;
+    // clear tracker for the delete button
+    self.deleteItem = nil;
+    
+    return itemCell;
 }
 
 #pragma mark - Cell Data
@@ -113,7 +119,7 @@
     self.groupName.tag = tag;
     self.groupIndicator.tag = tag;
     self.deleteAllButton.tag = tag;
-    self.tableView.tag = tag;
+    self.itemsView.tag = tag;
     
     // prepare container for print jobs list
     self.listPrintJobs = [NSMutableArray array];
@@ -147,7 +153,7 @@
 
 - (void)removePrintJob:(NSIndexPath*)indexPath
 {
-    [self.tableView deleteRowsAtIndexPaths:@[indexPath] withRowAnimation:UITableViewRowAnimationFade];
+    [self.itemsView deleteRowsAtIndexPaths:@[indexPath] withRowAnimation:UITableViewRowAnimationFade];
     [self reloadContents];
 }
 
@@ -155,30 +161,52 @@
 
 - (void)putDeleteButton:(UIGestureRecognizer*)gesture handledBy:(id<PrintJobHistoryGroupCellDelegate>)receiver usingActionOnTap:(SEL)actionOnTap
 {
-    // get the row swiped
-    CGPoint swipedItem = [gesture locationInView:self.tableView];
-    NSIndexPath* itemIndexPath = [self.tableView indexPathForRowAtPoint:swipedItem];
+    // get the specific item swiped
+    CGPoint swipedItem = [gesture locationInView:self.itemsView];
+    NSIndexPath* itemIndexPath = [self.itemsView indexPathForRowAtPoint:swipedItem];
     if (itemIndexPath != nil)
     {
         NSLog(@"[INFO][PrintJobCell] swiped left on item=%ld", (long)itemIndexPath.row);
         
-        UITableViewCell* cell = [self.tableView cellForRowAtIndexPath:itemIndexPath];
+        // check if there is already an item with a delete button
+        if ((self.deleteItem != nil) && (self.deleteItem.row != itemIndexPath.row))
+        {
+            NSLog(@"[INFO][PrintJobCell] canceling delete button for item=%ld", (long)self.deleteItem.row);
+            UITableViewCell* itemCell = [self.itemsView cellForRowAtIndexPath:self.deleteItem];
+            itemCell.accessoryView = nil;
+            itemCell.detailTextLabel.hidden = NO;
+        }
+        self.deleteItem = itemIndexPath;
+        
+        // check if this item already has a delete button
+        UITableViewCell* itemCell = [self.itemsView cellForRowAtIndexPath:itemIndexPath];
+        if (itemCell.accessoryView != nil)
+        {
+            NSLog(@"[INFO][PrintJobCell] already has delete button, ignore");
+            return;
+        }
+        
+        // create the delete button
         UIButton* deleteButton = [UIButton buttonWithType:UIButtonTypeCustom];
         [deleteButton setTitle:@"DELETE" forState:UIControlStateNormal];
         [deleteButton setTitleEdgeInsets:UIEdgeInsetsMake(10.0f, 15.0f, 10.0f, 15.0f)];
         [deleteButton setBackgroundColor:[UIColor redThemeColor]];
         [deleteButton setUserInteractionEnabled:YES];
         deleteButton.titleLabel.font = [UIFont systemFontOfSize:13.0f];
-        deleteButton.tag = (self.tableView.tag * TAG_FACTOR) + itemIndexPath.row; //<group>00<row>
+        deleteButton.tag = (self.itemsView.tag * TAG_FACTOR) + itemIndexPath.row; //<group>00<row>
         deleteButton.frame = CGRectMake(self.deleteAllButton.frame.origin.x,
                                         0,
                                         self.deleteAllButton.frame.size.width-15.0f,
-                                        cell.frame.size.height-10.0f);
+                                        itemCell.frame.size.height-10.0f);
+        
+        // set the handler for the tap action
         [deleteButton addTarget:receiver
                          action:actionOnTap
                forControlEvents:UIControlEventTouchUpInside];
-        cell.accessoryView = deleteButton;
-        cell.detailTextLabel.hidden = YES;
+        
+        // add to the view
+        itemCell.accessoryView = deleteButton;
+        itemCell.detailTextLabel.hidden = YES;
     }
     else
     {
@@ -188,9 +216,19 @@
     }
 }
 
+- (void)removeDeleteButton
+{
+    NSLog(@"[INFO][PrintJobCell] canceling delete button for item=%ld", (long)self.deleteItem.row);
+    
+    UITableViewCell* itemCell = [self.itemsView cellForRowAtIndexPath:self.deleteItem];
+    itemCell.accessoryView = nil;
+    itemCell.detailTextLabel.hidden = NO;
+    self.deleteItem = nil;
+}
+
 - (void)reloadContents
 {
-    [self.tableView reloadData];
+    [self.itemsView reloadData];
 }
 
 @end
